Fix compiler warnings: reorder, shadow, sign-compare, unused, and fallthrough
- Reordered member initializations to fix [-Wreorder]

- Annotated unused and set-but-unused variables/functions/parameters with [[maybe_unused]]

- Fixed variable shadowing to address [-Wshadow]

- Resolved signed/unsigned comparison issues ([-Wsign-compare], [-Wtype-limits])

- Added missing break; statements to eliminate [-Wimplicit-fallthrough=] warnings

- Refactored range-based loops for [-Wrange-loop-construct]

+++ avogadro/qtplugins/openbabel/openbabel.cpp
@@ -711,7 +711,7 @@ void OpenBabel::onGenerateConformersFinished(const QByteArray& output)
 
   // copy the coordinate sets
   m_molecule->clearCoordinate3d();
-  for (size_t i = 0; i < mol.coordinate3dCount(); ++i)
+  for (int i = 0; i < mol.coordinate3dCount(); ++i)
     m_molecule->setCoordinate3d(mol.coordinate3d(i), i);
 
   // energy data too
